added abstract deck class
